requirements.txt
Django==1.10
django-braces==1.9.0
django-oauth-toolkit==0.10.0
djangorestframework==3.4.6
extensions==0.4
gunicorn==19.6.0
oauthlib==1.0.3
psycopg2==2.6.2
six==1.10.0
Unipath==1.1
